refactor(exporter): consolidate map export pipeline into deep MapExporterEngine and delete shallow wrapper
