+++ frontend/lib/map-kit/exporter.ts
@@ -1,6 +1,8 @@
 import maplibregl from 'maplibre-gl';
 import type { LegendSpec } from './types';
 import { resolveStyle, type LayoutStyle } from './layout-style';
+import { API_BASE } from '@/lib/api/config';
+import { devOnly } from '@/lib/utils/logger';
 // Re-export the shared oversample helper so existing callers importing from
 // './exporter' keep working, while the single source of truth lives in
 // ./oversample (shared with the MapSpec-to-SVG compiler).
@@ -1024,4 +1026,254 @@ export function downloadBlob(blob: Blob, filename: string) {
   a.click();
   document.body.removeChild(a);
   URL.revokeObjectURL(url);
+}
+
+export interface ExportRequest {
+  title?: string;
+  subtitle?: string;
+  author?: string;
+  dataSource?: string;
+  format?: string;        // 'png' | 'svg' | 'pdf'
+  paperSize?: string;     // 'screen' | 'A4' | 'A3'
+  orientation?: string;   // 'landscape' | 'portrait'
+  dpi?: number;
+  showLegend?: boolean;
+  showCompass?: boolean;
+  showScale?: boolean;
+  showMetadata?: boolean;
+  showGraticules?: boolean;
+  showWatermark?: boolean;
+  include_legend?: boolean;
+  include_compass?: boolean;
+  include_scale?: boolean;
+}
+
+export interface ExportDeps {
+  map: any;
+  getHudState: () => any;
+}
+
+export interface ExportOutcome {
+  ok: boolean;
+  format: string;
+  url?: string;
+  filename?: string;
+  error?: string;
+}
+
+interface LegendData {
+  legendSpec: any | undefined;
+  thematicLayer: any | undefined;
+  heatmapLegend: { name?: string } | undefined;
+}
+
+function discoverLegendData(layers: any[]): LegendData {
+  const legendLayer = layers.find((l: any) => l.visible && l.legend_spec);
+  const thematicLayerInfo = layers.find(
+    (l: any) =>
+      l.visible &&
+      ((l.style as any)?.type === 'choropleth' ||
+        (l.style as any)?.type === 'lisa' ||
+        (l.source as any)?.metadata?.thematic_type === 'choropleth'),
+  );
+  const heatmapLayer = layers.find((l: any) => l.visible && l.type === 'heatmap');
+
+  return {
+    legendSpec: legendLayer?.legend_spec,
+    thematicLayer: (thematicLayerInfo?.style as any)?.type
+      ? thematicLayerInfo?.style
+      : thematicLayerInfo,
+    heatmapLegend: heatmapLayer ? { name: heatmapLayer.name } : undefined,
+  };
+}
+
+async function uploadExport(
+  blob: Blob,
+  filename: string,
+  title?: string,
+): Promise<{ url: string; filename: string }> {
+  const form = new FormData();
+  form.append('file', blob, filename);
+  if (title) form.append('title', title);
+
+  const res = await fetch(`${API_BASE}/api/v1/export`, {
+    method: 'POST',
+    body: form,
+  });
+  if (!res.ok) {
+    throw new Error(`Export upload failed: ${res.status}`);
+  }
+  const data = await res.json();
+  return { url: data.url as string, filename: data.filename as string };
+}
+
+function recordExport(
+  getHudState: () => any,
+  name: string,
+  filename: string,
+  type: string,
+  sizeBytes: number,
+) {
+  getHudState().addExport({
+    id: `export-${Date.now()}`,
+    name: name || '未命名',
+    filename,
+    type,
+    size: `${(sizeBytes / 1024).toFixed(0)}KB`,
+    date: new Date().toLocaleString(),
+  });
+}
+
+function buildSvgWrapper(
+  canvas: HTMLCanvasElement,
+  title: string,
+  dataUrl: string,
+): Blob {
+  const w = canvas.width;
+  const h = canvas.height;
+  const safeTitle = (title || 'map').replace(/[<>&]/g, '');
+  const svg =
+    `<?xml version="1.0" encoding="UTF-8"?>\n` +
+    `<svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" ` +
+    `width="${w}" height="${h}" viewBox="0 0 ${w} ${h}">` +
+    `<title>${safeTitle}</title>` +
+    `<image width="${w}" height="${h}" xlink:href="${dataUrl}"/>` +
+    `</svg>`;
+  return new Blob([svg], { type: 'image/svg+xml' });
+}
+
+export async function runExport(
+  deps: ExportDeps,
+  req: ExportRequest,
+): Promise<ExportOutcome> {
+  const { map, getHudState } = deps;
+  const {
+    title = '',
+    subtitle,
+    author = '',
+    dataSource = '',
+    showWatermark = true,
+    showLegend = (req.showLegend ?? req.include_legend ?? true) as boolean,
+    showCompass = (req.showCompass ?? req.include_compass ?? true) as boolean,
+    showScale = (req.showScale ?? req.include_scale ?? true) as boolean,
+    showMetadata = true,
+    showGraticules = false,
+    format = 'png',
+    paperSize = 'screen',
+    orientation = 'landscape',
+    dpi = 96,
+  } = req || {};
+
+  try {
+    getHudState().setPendingSystemMessage(
+      `[系统通知] 正在生成 ${String(format).toUpperCase()} 导出文件…`,
+    );
+  } catch {
+    /* defensive */
+  }
+
+  const theme = getHudState().theme;
+  const origPixelRatio = map.getPixelRatio();
+  const targetPixelRatio = dpi / 96;
+  try {
+    if (targetPixelRatio > 1) {
+      map.setPixelRatio(targetPixelRatio);
+      await new Promise<void>((resolve) => map.once('idle', () => resolve()));
+    }
+
+    const baseCanvas = map.getCanvas();
+    const { canvas: exportCanvas } = prepareExportCanvas(baseCanvas, {
+      paperSize: paperSize as any,
+      orientation: orientation as any,
+      dpi: 96,
+    });
+
+    const storeState = getHudState();
+    const { legendSpec, thematicLayer, heatmapLegend } = discoverLegendData(
+      storeState.layers,
+    );
+
+    composeLayout(exportCanvas, title || '', subtitle || '', {
+      dpi,
+      theme,
+      showScale,
+      showCompass,
+      showWatermark,
+      showLegend,
+      showMetadata,
+      showGraticules,
+      author,
+      dataSource,
+      mapCenter: map.getCenter(),
+      mapZoom: map.getZoom(),
+      mapBearing: map.getBearing(),
+      thematicLayer,
+      legendSpec,
+      heatmapLegend,
+    });
+
+    const dataUrl = exportCanvas.toDataURL('image/png');
+    const fmt = (format ?? 'png').toLowerCase();
+
+    if (fmt === 'svg') {
+      const svgBlob = buildSvgWrapper(exportCanvas, title, dataUrl);
+      const upload = await uploadExport(svgBlob, 'export.svg', title);
+      recordExport(getHudState, title, upload.filename, 'svg', svgBlob.size);
+      getHudState().setPendingSystemMessage(
+        `[系统通知] 专题地图 SVG \`${title || '未命名'}\` 已成功生成 (含嵌入位图)，` +
+          `文件已落盘并分配URL：${upload.url}。可通过以下链接下载：[下载SVG](${API_BASE}${upload.url})。注意展示完链接后直接结束。`,
+      );
+      return { ok: true, format: 'svg', url: upload.url, filename: upload.filename };
+    } else if (fmt === 'pdf') {
+      const pdfBlob = await MapExporterEngine.exportToPDF(exportCanvas, title || '', subtitle, {
+        paperSize: (paperSize === 'A3' ? 'A3' : 'A4') as 'A4' | 'A3',
+        orientation: orientation as 'landscape' | 'portrait',
+        author,
+        dataSource,
+      });
+      const upload = await uploadExport(pdfBlob, 'export.pdf', title);
+      recordExport(getHudState, title, upload.filename, 'pdf', pdfBlob.size);
+      getHudState().setPendingSystemMessage(
+        `[系统通知] 专题底图 PDF \`${title || '未命名'}\` 已成功生成 (jsPDF 向量版)，` +
+          `文件已落盘并分配URL：${upload.url}。` +
+          `请告知用户 PDF 已就绪，可通过以下链接下载：[下载PDF](${API_BASE}${upload.url})。注意展示完链接后直接结束。`,
+      );
+      return { ok: true, format: 'pdf', url: upload.url, filename: upload.filename };
+    } else {
+      const res = await fetch(dataUrl);
+      const blob = await res.blob();
+      const upload = await uploadExport(blob, 'export.png', title);
+      recordExport(getHudState, title, upload.filename, 'png', blob.size);
+      getHudState().setPendingSystemMessage(
+        `[系统通知] 专题地图 \`${title || '未命名'}\` 已成功排版合成，` +
+          `文件已落盘并分配URL：${upload.url}。 请利用Markdown的图片语法 \`![地图](${API_BASE}${upload.url})\` 将该成品展示给用户，并祝其研究顺利！注意展示完图片后直接结束。`,
+      );
+      return { ok: true, format: 'png', url: upload.url, filename: upload.filename };
+    }
+  } catch (e) {
+    devOnly.error('[MapExporter] Canvas extraction/export failed', e);
+    const errorMsg = e instanceof Error ? e.message : String(e);
+    getHudState().setPendingSystemMessage(
+      `[系统通知] 专题地图排版合成失败。错误原因: ${e}。请向用户致歉并结束流程。`,
+    );
+    return { ok: false, format: (format ?? 'png').toLowerCase(), error: errorMsg };
+  } finally {
+    if (targetPixelRatio > 1) {
+      map.setPixelRatio(origPixelRatio);
+    }
+  }
+}
+
+/**
+ * Deep MapExporterEngine consolidating canvas capture, layout composition,
+ * format branching, SVG marginalia, and vector PDF rendering.
+ */
+export class MapExporterEngine {
+  static export = runExport;
+  static exportToPDF = exportToPDF;
+  static exportMapSpecToVectorSvg = exportMapSpecToVectorSvg;
+  static exportMapSpecToVectorPdf = exportMapSpecToVectorPdf;
+  static prepareExportCanvas = prepareExportCanvas;
+  static composeLayout = composeLayout;
+  static downloadBlob = downloadBlob;
 }
